瀑布流显示图片，其中使用Glide不能设置其图片质量，设置向上滑的动作的动画，加载过快，图片有闪动

+++ app/src/main/java/com/demo/panguso/mvp_mode/mvp/view/RationImageView.java
@@ -32,7 +32,7 @@ public class RationImageView extends ImageView {
     @Override
     protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
         if (originalWidth > 0 && originalHeight > 0) {
-            float ratio = originalWidth / originalHeight;
+            float ratio = (float) originalWidth / (float)originalHeight;
             int width = MeasureSpec.getSize(widthMeasureSpec);
             int height = MeasureSpec.getSize(heightMeasureSpec);
             if (width > 0) {
